third_party/WebKit/LayoutTests/http/tests/security/no-indexeddb-from-sandbox.html

<script>
if (window.testRunner)
    testRunner.dumpAsText();
</script>
<iframe sandbox="allow-scripts"
    src="data:text/html,
         <script>
             try {
                 indexedDB.open('test');
                 alert('FAIL: indexedDB.open() should throw a SECURITY_ERR in a sandbox.');
             } catch (e) {
                 if (e.code === DOMException.SECURITY_ERR)
                     alert('PASS: indexedDB.open() threw a SECURITY_ERR!');
                 else
                     alert('FAIL: indexedDB.open() threw a ' + e.name);
             }
         </script>"
     ></iframe>
