third_party/WebKit/LayoutTests/http/tests/security/no-indexeddb-from-sandbox-expected.txt

ALERT: PASS: indexedDB.open() threw a SECURITY_ERR!
